@@ -1,132 +1,123 @@
 {
-    "name": "react-responsive-carousel",
-    "version": "3.2.22",
-    "description": "React Responsive Carousel",
-    "author": {
-        "name": "Leandro Augusto Lemos",
-        "url": "http://leandrowd.github.io/"
-    },
-    "main": "lib/js/index.js",
-    "types": "lib/ts/index.d.ts",
-    "license": "MIT",
-    "keywords": [
-        "react",
-        "carousel",
-        "gallery",
-        "image-gallery",
-        "slider",
-        "responsive",
-        "swipe",
-        "mobile-friendly",
-        "react-component",
-        "view"
+  "name": "react-sliding-carousel",
+  "version": "1.0.0",
+  "description": "React Sliding Carousel",
+  "author": {
+    "name": "Luka Mandić",
+    "url": "https://github.com/mandicOfisl"
+  },
+  "main": "lib/js/index.js",
+  "types": "lib/ts/index.d.ts",
+  "license": "MIT",
+  "keywords": [
+    "react",
+    "carousel",
+    "gallery",
+    "image-gallery",
+    "slider",
+    "responsive",
+    "swipe",
+    "mobile-friendly",
+    "react-component",
+    "view"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mandicOfisl/react-sliding-carousel.git"
+  },
+  "bugs": {
+    "url": "https://github.com/mandicOfisl/react-sliding-carousel/issues"
+  },
+  "homepage": "https://github.com/mandicOfisl/react-sliding-carousel",
+  "scripts": {
+    "start": "parcel src/index.html src/assets/**/*",
+    "storybook": "start-storybook -p 9001 -s ./src -c .storybook",
+    "changelog": "auto-changelog --ignore-commit-pattern=\"(Merge pull request|Merge branch|Updating changelog|Prepare for publishing)\" --breaking-pattern \"BREAKING CHANGE:\"",
+    "format": "prettier \"**/*.{js,ts,tsx,json}\"",
+    "format:check": "yarn format --check",
+    "format:write": "yarn format --write",
+    "test": "yarn format:check && yarn typecheck && yarn jest && yarn jest-ssr",
+    "jest": "jest",
+    "jest-ssr": "jest --testEnvironment=node ./src/__tests__/SSR.tsx",
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "update-snapshots": "jest --updateSnapshot",
+    "build": "yarn lib:build",
+    "lib:build": "yarn lib:build-js && yarn lib:build-styles && yarn lib:build-types",
+    "lib:build-js": "babel ./src -d lib/js --ignore './src/__tests__' --extensions .ts,.tsx --config-file ./babel.config.json",
+    "lib:build-styles": "mkdirp lib/styles && node-sass src/carousel.scss > lib/styles/carousel.css && node-sass --output-style compressed src/carousel.scss > lib/styles/carousel.min.css",
+    "lib:build-types": "tsc -p tsconfig.types.json",
+    "lib:pre-publish": "npm version patch && git push origin master",
+    "lib:publish": "npm publish && git push --tags",
+    "lib:post-publish": "yarn changelog && yarn update-codesandbox && git add . && git commit -m 'Updating changelog and codesandbox' && git push origin master",
+    "lib:build-and-publish": "yarn lib:build && yarn lib:pre-publish && npm run lib:publish && yarn lib:post-publish",
+    "prepublish-to-npm": "git pull && yarn build",
+    "prepublish-prerelease-to-npm": "git pull && yarn build && git add . && git commit -m 'Prepare for publishing prerelease'",
+    "publish-prerelease-to-npm": "(git pull && npm version prerelease --preid=next && npm publish)"
+  },
+  "devDependencies": {
+    "@babel/cli": "^7.26.4",
+    "@babel/core": "^7.26.8",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/preset-env": "^7.26.8",
+    "@babel/preset-react": "^7.26.3",
+    "@babel/preset-typescript": "^7.26.0",
+    "@kadira/react-storybook-addon-info": "^3.4.0",
+    "@kadira/storybook": "^2.35.3",
+    "@storybook/addon-actions": "^8.5.3",
+    "@storybook/addon-essentials": "^8.5.3",
+    "@storybook/addon-knobs": "^8.0.1",
+    "@storybook/addon-storysource": "^8.5.3",
+    "@storybook/addon-viewport": "^8.5.3",
+    "@storybook/react": "^8.5.3",
+    "@types/enzyme": "^3.10.18",
+    "@types/jest": "^29.5.14",
+    "@types/react": "^19.0.8",
+    "@types/react-dom": "^19.0.3",
+    "@types/react-test-renderer": "^19.0.0",
+    "auto-changelog": "^2.5.0",
+    "babel-loader": "^9.2.1",
+    "css-loader": "^7.1.2",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.8",
+    "gh-pages": "^6.3.0",
+    "husky": "^9.1.7",
+    "jest-cli": "^29.7.0",
+    "mkdirp": "^3.0.1",
+    "node-sass": "^9.0.0",
+    "parcel-bundler": "^1.12.5",
+    "prettier": "^3.5.0",
+    "pretty-quick": "^4.0.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-player": "^2.16.0",
+    "react-test-renderer": "^19.0.0",
+    "sass-loader": "^16.0.4",
+    "style-loader": "^4.0.0",
+    "typescript": "^5.7.3"
+  },
+  "dependencies": {
+    "classnames": "^2.5.1",
+    "prop-types": "^15.8.1",
+    "react-easy-swipe": "^0.0.23"
+  },
+  "jest": {
+    "unmockedModulePathPatterns": [
+      "node_modules"
     ],
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/leandrowd/react-responsive-carousel.git"
-    },
-    "bugs": {
-        "url": "https://github.com/leandrowd/react-responsive-carousel/issues"
-    },
-    "homepage": "http://leandrowd.github.io/react-responsive-carousel/",
-    "scripts": {
-        "start": "parcel src/index.html src/assets/**/*",
-        "storybook": "start-storybook -p 9001 -s ./src -c .storybook",
-        "changelog": "auto-changelog --ignore-commit-pattern=\"(Merge pull request|Merge branch|Updating changelog|Prepare for publishing)\" --breaking-pattern \"BREAKING CHANGE:\"",
-        "update-codesandbox": "cd codesandbox/default && yarn add react-responsive-carousel@latest",
-        "format": "prettier \"**/*.{js,ts,tsx,json}\"",
-        "format:check": "yarn format --check",
-        "format:write": "yarn format --write",
-        "test": "yarn format:check && yarn typecheck && yarn jest && yarn jest-ssr",
-        "jest": "jest",
-        "jest-ssr": "jest --testEnvironment=node ./src/__tests__/SSR.tsx",
-        "typecheck": "tsc -p tsconfig.json --noEmit",
-        "update-snapshots": "jest --updateSnapshot",
-        "build": "yarn lib:build",
-        "lib:build": "yarn lib:build-js && yarn lib:build-styles && yarn lib:build-types",
-        "lib:build-js": "babel ./src -d lib/js --ignore './src/__tests__' --extensions .ts,.tsx --config-file ./babel.config.json",
-        "lib:build-styles": "mkdirp lib/styles && node-sass src/carousel.scss > lib/styles/carousel.css && node-sass --output-style compressed src/carousel.scss > lib/styles/carousel.min.css",
-        "lib:build-types": "tsc -p tsconfig.types.json",
-        "lib:pre-publish": "npm version patch && git push origin master",
-        "lib:publish": "npm publish && git push --tags",
-        "lib:post-publish": "yarn changelog && yarn update-codesandbox && git add . && git commit -m 'Updating changelog and codesandbox' && git push origin master",
-        "lib:build-and-publish": "yarn lib:build && yarn lib:pre-publish && npm run lib:publish && yarn lib:post-publish",
-        "website:build": "parcel build ./src/index.html --out-dir temp/website",
-        "website:copy-assets": "cp -r ./src/assets temp/website/assets && cp -r ./CNAME temp/website/CNAME",
-        "website:storybook": "build-storybook -s ./src -o ./temp/website/storybook",
-        "website:deploy": "gh-pages -d temp/website",
-        "website:create-and-publish": "yarn website:build && yarn website:copy-assets && yarn website:storybook && yarn website:deploy",
-        "prepublish-to-npm": "git pull && yarn build",
-        "publish-to-npm": "npm version patch && npm publish && git push --tags",
-        "postpublish-to-npm": "yarn changelog && yarn update-codesandbox && git add . && git commit -m 'Updating changelog and codesandbox' && git push origin master",
-        "prepublish-prerelease-to-npm": "git pull && yarn build && git add . && git commit -m 'Prepare for publishing prerelease'",
-        "publish-prerelease-to-npm": "(git pull && npm version prerelease --preid=next && npm publish)"
-    },
-    "devDependencies": {
-        "@babel/cli": "^7.8.4",
-        "@babel/core": "^7.9.0",
-        "@babel/plugin-proposal-class-properties": "^7.8.3",
-        "@babel/preset-env": "^7.9.5",
-        "@babel/preset-react": "^7.9.4",
-        "@babel/preset-typescript": "^7.9.0",
-        "@kadira/react-storybook-addon-info": "^3.4.0",
-        "@kadira/storybook": "^2.35.3",
-        "@storybook/addon-actions": "^5.3.18",
-        "@storybook/addon-essentials": "^5.3.18",
-        "@storybook/addon-knobs": "^5.3.18",
-        "@storybook/addon-storysource": "^5.3.18",
-        "@storybook/addon-viewport": "^5.3.18",
-        "@storybook/react": "^5.3.18",
-        "@types/classnames": "^2.2.10",
-        "@types/enzyme": "^3.10.5",
-        "@types/jest": "^25.2.1",
-        "@types/react": "^16.9.34",
-        "@types/react-dom": "^16.9.6",
-        "@types/react-test-renderer": "^16.9.2",
-        "auto-changelog": "^1.10.2",
-        "babel-loader": "^8.1.0",
-        "css-loader": "^3.5.2",
-        "enzyme": "^3.11.0",
-        "enzyme-adapter-react-16": "^1.15.2",
-        "gh-pages": "^2.2.0",
-        "husky": "^3.0.9",
-        "jest-cli": "^25.3.0",
-        "mkdirp": "^1.0.4",
-        "node-sass": "^4.13.1",
-        "parcel-bundler": "^1.12.4",
-        "prettier": "^1.18.2",
-        "pretty-quick": "^2.0.0",
-        "react": "^16.9.0",
-        "react-dom": "^16.9.0",
-        "react-player": "^1.15.3",
-        "react-test-renderer": "^16.9.0",
-        "sass-loader": "^8.0.2",
-        "style-loader": "^1.1.3",
-        "typescript": "^3.8.3"
-    },
-    "dependencies": {
-        "classnames": "^2.2.5",
-        "prop-types": "^15.5.8",
-        "react-easy-swipe": "^0.0.21"
-    },
-    "jest": {
-        "unmockedModulePathPatterns": [
-            "node_modules"
-        ],
-        "rootDir": "src",
-        "setupFilesAfterEnv": [
-            "../setupTests.js"
-        ]
-    },
-    "husky": {
-        "hooks": {
-            "pre-commit": "pretty-quick --staged"
-        }
-    },
-    "auto-changelog": {
-        "output": "CHANGELOG.md",
-        "template": "keepachangelog",
-        "unreleased": true,
-        "commitLimit": true
+    "rootDir": "src",
+    "setupFilesAfterEnv": [
+      "../setupTests.js"
+    ]
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "pretty-quick --staged"
     }
+  },
+  "auto-changelog": {
+    "output": "CHANGELOG.md",
+    "template": "keepachangelog",
+    "unreleased": true,
+    "commitLimit": true
+  }
 }
